pyproject.toml
[tool.poetry]
name = "user-management-service"
version = "0.1.0"
description = ""
authors = ["Elizaveta Krutsko <elizaveta.krutsko@innowise-group.com>"]
readme = "README.md"
packages = [{include = "user_management_service"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.98.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
gunicorn = "^20.1.0"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.17"}
pydantic = {extras = ["dotenv"], version = "^1.10.9"}
aioredis = "^2.0.1"
sqlalchemy-imageattach = "^1.1.0"
psycopg2-binary = "^2.9.6"
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
